+++ packages/present/package.json
@@ -1,0 +1,71 @@
+{
+  "name": "@hopae/sd-jwt-present",
+  "version": "2.0.1",
+  "description": "sd-jwt draft 7 implementation in typescript",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    }
+  },
+  "scripts": {
+    "build": "rm -rf **/dist && tsup",
+    "test": "vitest run ./src/test/*.spec.ts",
+    "test:browser": "vitest run ./src/test/*.spec.ts --environment jsdom",
+    "test:cov": "vitest run --coverage"
+  },
+  "keywords": [
+    "verifiable-credentials",
+    "digital-wallet",
+    "sd-jwt",
+    "sdjwt",
+    "sd-jwt-vc",
+    "decentralized-identity",
+    "ssi",
+    "oauth",
+    "oauth2",
+    "openid-connect"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/openwallet-foundation-labs/sd-jwt-js"
+  },
+  "author": "Lukas.J.Han <lukas.j.han@gmail.com>",
+  "homepage": "https://github.com/openwallet-foundation-labs/sd-jwt-js/wiki",
+  "bugs": {
+    "url": "https://github.com/openwallet-foundation-labs/sd-jwt-js/issues"
+  },
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "@types/node": "^20.10.2",
+    "@vitest/coverage-v8": "^1.2.2",
+    "jsdom": "^24.0.0",
+    "lerna": "^8.1.2",
+    "ts-node": "^10.9.1",
+    "tsup": "^8.0.2",
+    "typescript": "^5.3.2",
+    "vite": "^5.1.1",
+    "vitest": "^1.2.2"
+  },
+  "dependencies": {
+    "@hopae/sd-jwt-util": "workspace:*",
+    "@hopae/sd-jwt-type": "workspace:*",
+    "@hopae/sd-jwt-decode": "workspace:*"
+  },
+  "tsup": {
+    "entry": [
+      "./src/index.ts"
+    ],
+    "sourceMap": true,
+    "splitting": false,
+    "clean": true,
+    "dts": true,
+    "format": [
+      "cjs",
+      "esm"
+    ]
+  }
+}
