+++ packages/present/src/test/present.spec.ts
@@ -1,0 +1,7 @@
+import { describe, expect, test } from 'vitest';
+
+describe('Error tests', () => {
+  test('Detail', () => {
+    expect(1).toBe(1);
+  });
+});
